+++ central/sensor/service/connection/manager.go
@@ -21,6 +21,7 @@ type Manager interface {
 		baselineMgr common.ProcessBaselineManager,
 		networkBaselineMgr common.NetworkBaselineManager,
 		delegatedRegistryConfigMgr common.DelegatedRegistryConfigManager,
+		imageIntegrationMgr common.ImageIntegrationManager,
 		autoTriggerUpgrades *concurrency.Flag) error
 
 	// Connection-related methods.